seeder_api/src/configs/guard/post-request/post-request.guard.ts
import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
import { Observable } from 'rxjs';

@Injectable()
export class PostRequestGuard implements CanActivate {
  canActivate(
    context: ExecutionContext,
  ): boolean | Promise<boolean> | Observable<boolean> {
    const req = context.switchToHttp().getRequest();
    const user = req.user;
    const body = req.body;
    body.active = true;
    body.createdBy = user.id;
    body.createTimestamp = new Date();
    return true;
  }
}
